+++ projects/angular2-draggable/src/lib/widgets/resize-handle.ts
@@ -28,8 +28,8 @@ export class ResizeHandle {
 
     // create and register event listener
     this._onResize = (event) => { onMouseDown(event, this); };
-    handle.addEventListener('mousedown', this._onResize);
-    handle.addEventListener('touchstart', this._onResize);
+    handle.addEventListener('mousedown', this._onResize, { passive: false });
+    handle.addEventListener('touchstart', this._onResize, { passive: false });
 
     // done
     this._handle = handle;