+++ src/app/views/usage/options/options.component.ts
@@ -17,6 +17,7 @@ export class OptionsComponent implements OnInit {
   preventDefaultEvent = false;
   trackPosition = true;
   position;
+  lockAxis;
 
   constructor() { }