+++ src/assets/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.2.0 (2018-12-22)
+
+#### New
++ **ngDraggable**: add [lockAxis] input to restrict dragging to a specific axis by locking another one.
+
+#### Bugfix
++ **ngDraggable**:
+    + fix [issue #112](https://github.com/xieziyu/angular2-draggable/issues/112): Control change detection with HostListener events. Performance updated.
+    + fix [issue #128](https://github.com/xieziyu/angular2-draggable/issues/128): Multiple phone draggables at the same time.
++ **ngResizable**:
+    + fix [issue #132](https://github.com/xieziyu/angular2-draggable/issues/132): Aspect ratio feature exits Y-Axis boundary on resize.
+    + Performance updated.
+---
+
 ## 2.1.9 (2018-11-29)
 
 #### Bugfix